Pravaha/backend/requirements.txt
# ── Core Framework ────────────────────────────────────────────────────
fastapi==0.110.2
uvicorn==0.29.0
starlette==0.37.2
python-multipart==0.0.9
pydantic==2.6.4
email_validator==2.1.1

# ── Auth & Security ──────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib==1.7.4
itsdangerous==2.1.2

# ── Database ─────────────────────────────────────────────────────────
pymongo==4.7.0
dnspython==2.6.1
bson>=0.5.10

# ── LLM / AI ────────────────────────────────────────────────────────
langchain==0.1.13
langchain-core==0.1.46
langchain-community==0.0.29
langchain-groq==0.1.3
langchain-cohere==0.1.4
langchain-text-splitters==0.0.1
langsmith==0.1.38
groq==0.9.0

# ── Vector DB & Embeddings ──────────────────────────────────────────
pinecone-client==3.2.2
cohere==5.3.3

# ── Document Processing ─────────────────────────────────────────────
pypdf==4.2.0
pypandoc==1.13
markdown==3.6

# ── HTTP / Networking ───────────────────────────────────────────────
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.3

# ── Environment ─────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Utilities ───────────────────────────────────────────────────────
tiktoken==0.6.0
tenacity==8.2.3
orjson==3.10.0
PyYAML==6.0.1
